+++ src/handlers/SubtasksHandler.java
@@ -2,8 +2,7 @@ package handlers;
 
 import com.google.gson.JsonSyntaxException;
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
-import controllers.InMemoryTaskManager;
+import controllers.TaskManager;
 import enums.Endpoint;
 import enums.Status;
 import model.SubTask;
@@ -15,40 +14,33 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Optional;
 
-public class SubtasksHandler extends BaseHandler implements HttpHandler {
+public class SubtasksHandler extends BaseHandler {
 
-    private final InMemoryTaskManager taskManager;
+    private final TaskManager taskManager;
 
-    public SubtasksHandler(InMemoryTaskManager taskManager) {
+    public SubtasksHandler(TaskManager taskManager) {
         this.taskManager = taskManager;
     }
 
-    @Override
-    public void handle(HttpExchange exchange) throws IOException {
-        Endpoint endpoint = getEndpoint(exchange.getRequestURI().getPath(), exchange.getRequestMethod());
-        switch (endpoint) {
-            case GET_SUBTASKS: {
-                handleGetSubTasks(exchange);
-                break;
-            }
-            case GET_SUBTASK_BY_ID: {
-                handleGetSubTaskById(exchange);
-                break;
-            }
-            case DELETE_SUBTASK: {
-                handleDeleteSubTask(exchange);
-                break;
-            }
-            case POST_UPDATE_SUBTASK: {
-                handleUpdateSubTask(exchange);
-                break;
-            }
-            case POST_ADD_SUBTASK: {
-                handleAddSubTask(exchange);
-                break;
-            }
-            default:
-                sendNotFound(exchange, "Такого эндпоинта не существует");
+    protected void processGet(HttpExchange exchange, Endpoint endpoint) throws IOException {
+        if (endpoint.equals(Endpoint.GET_SUBTASKS)) {
+            handleGetSubTasks(exchange);
+        } else if (endpoint.equals(Endpoint.GET_SUBTASK_BY_ID)) {
+            handleGetSubTaskById(exchange);
+        }
+    }
+
+    protected void processPost(HttpExchange exchange, Endpoint endpoint) throws IOException {
+        if (endpoint.equals(Endpoint.POST_ADD_SUBTASK)) {
+            handleAddSubTask(exchange);
+        } else if (endpoint.equals(Endpoint.POST_UPDATE_SUBTASK)) {
+            handleUpdateSubTask(exchange);
+        }
+    }
+
+    protected void processDelete(HttpExchange exchange, Endpoint endpoint) throws IOException {
+        if (endpoint.equals(Endpoint.DELETE_SUBTASK)) {
+            handleDeleteSubTask(exchange);
         }
     }
